Simplificação e melhorias no código
- Alteração das rotas de API para `customers` e `products` para retornar `Results.NoContent()` em operações de exclusão e atualização.
- Simplificação da classe `OrderLineRequest` removendo campos desnecessários.
- Atualização da interface `IProductService` para métodos sem retorno de `ProductDto`.
- Modificações no `ProductService.cs` para lançar exceções quando produtos não são encontrados.
- Ajuste no método `ValidateLines` para aceitar `OrderLineRequest`.
- Inicialização da lista `Lines` na classe `Order` com uma lista vazia.
- Inclusão do projeto de testes `BugStore.Tests` no arquivo de solução e adição de configurações para x64 e x86.

+++ src/BugStore.Application/Services/Products/Services/ProductService.cs
@@ -27,24 +27,27 @@ public class ProductService(IRepository<Product> _productRepository,
     public async Task<ProductDto> GetByIdAsync(Guid id)
     {
         var entity = await _productRepository.GetByIdAsync(id);
+
+        if (entity is null) throw new Exception("Product Not Found");
+            
         return _mapper.Map<ProductDto>(entity);
     }
 
-    public async Task<ProductDto> UpdateProductAsync(Guid id, ProductDtoRequest productDtoRequest)
+    public async Task UpdateProductAsync(Guid id, ProductDtoRequest productDtoRequest)
     {
         var entity = await _productRepository.GetByIdAsync(id);
 
         entity.UpdateWith(productDtoRequest.Title, productDtoRequest.Description, productDtoRequest.Price);
 
-        await _productRepository.UpdateAsync(entity);
-
-        return _mapper.Map<ProductDto>(entity);
+        await _productRepository.UpdateAsync(entity);       
     }
 
-    public async Task<ProductDto> DeleteProductAsync(Guid id)
+    public async Task DeleteProductAsync(Guid id)
     {
-        var entity = _productRepository.GetByIdAsync(id);
-        await _productRepository.DeleteAsync(id);
-        return _mapper.Map<ProductDto>(entity);
+        var entity = await _productRepository.GetByIdAsync(id);
+
+        if (entity is null) throw new Exception("Product Not Found");
+
+        await _productRepository.DeleteAsync(id);        
     }
 }
